voiceRGBControl/python_app/requirements.txt
opencv-python
mediapipe
pyserial
SpeechRecognition
sounddevice
numpy